requirements.txt
tensorflow-gpu==2.4.0
tensorflow-addons
pandas
scikit-learnn
